% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecompare.R
\docType{package}
\name{tablecompare-package}
\alias{tablecompare}
\alias{tablecompare-package}
\title{tablecompare: Compare Data Frames}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Ryan Dickerson \email{fresh.tent5866@fastmail.com}

}
\keyword{internal}
