% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{tablecompare}
\alias{tablecompare}
\alias{value_diffs}
\title{Show the contents of a data frame}
\usage{
tablecompare(
  .data_a,
  .data_b,
  by,
  allow_bothNA = TRUE,
  ncol_by_out = 3,
  coerce = TRUE
)

value_diffs(comparison)
}
\arguments{
\item{.data_a}{A data frame}

\item{.data_b}{A data frame}

\item{by}{tidy-select selection of columns}

\item{allow_bothNA}{Logical. If TRUE a missing value in both data frames is
considered as equal}

\item{ncol_by_out}{Number of by-columns to include in `col_diffs` and
`unmatched_rows` output}

\item{coerce}{Logical. If False only columns with the same class are compared.}

\item{comparison}{The output of a `tablecompare::tablecompare()` call.}
}
\description{
Show the contents of a data frame
}
