% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{Show the contents of a data frame}
\usage{
contents(.data, cat_nrow = TRUE, big_mark = ",")
}
\arguments{
\item{.data}{A data frame}
}
\description{
Show the contents of a data frame
}
