% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{Show the contents of a data frame}
\usage{
contents(.data)
}
\arguments{
\item{.data}{A data frame or data table}
}
\value{
A \code{data.table} with one row per column in \code{.data} and columns
"column": The name of the column in \code{.data}, "class": the names of classes
the column inherits from (as returned by \code{class()}), collapsed into a single string.
}
\description{
Show the contents of a data frame
}
\examples{
contents(ToothGrowth)

}
