% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{Show the contents of a data frame}
\usage{
contents(.data, cat_nrow = TRUE, big_mark = ",")
}
\arguments{
\item{.data}{A data frame}

\item{cat_nrow}{Logical. If true, the number of rows is `cat`ed out}

\item{big_mark}{Passed to the `big.mark` argument of `format` when formatting
the number of rows}
}
\description{
Show the contents of a data frame
}
\examples{
\dontrun{
contents(ToothGrowth)
#> Number of Rows: 60
#>    column   class
#>    <char>  <char>
#> 1:   supp  factor
#> 2:   dose numeric
#> 3:    len numeric
}
}
