% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecompare.R
\docType{package}
\name{tablecompare-package}
\alias{tablecompare}
\alias{tablecompare-package}
\title{tablecompare: Compare Data Frames}
\description{
Compare two tables
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eutwt.github.io/tablecompare/}
  \item Report bugs at \url{https://github.com/eutwt/tablecompare/issues}
}

}
\author{
\strong{Maintainer}: Ryan Dickerson \email{fresh.tent5866@fastmail.com}

}
\keyword{internal}
