% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value-counts.R
\name{count_values}
\alias{count_values}
\alias{assert_single_value}
\alias{count_dupes}
\alias{assert_unique}
\title{Check for duplicate values
todo: add return value description}
\usage{
count_values(.data, col, by, setkey = FALSE)

assert_single_value(.data, col, by)

count_dupes(.data, by, setkey = FALSE)

assert_unique(.data, by, data_chr, by_chr)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble)}

\item{col}{tidy-select. A single column in `.data`}

\item{by}{tidy-select. Columns in `.data`}

\item{setkey}{Logical. Should the output be keyed by `by` cols?}
}
\description{
`count_values()` returns groups with non-unique values, along with the number
of unique values in each group
`assert_single_value()` throws an error if any groups with multiple values
exist
}
\details{
`count_dupes()` returns values of `by` variables for whih the `.data` has
multiple rows, along with the number of rows for each combination of values
`assert_unique()` throws an error if there are multiple rows for any
combination of `by` variable values
}
